Fix media player time type

+++ include/mediastreamer2/msinterfaces.h
@@ -135,7 +135,7 @@ typedef enum _MSPlayerState MSPlayerState;
 	MS_FILTER_METHOD(MSFilterPlayerInterface,7,int)
 
 #define MS_PLAYER_GET_CURRENT_POSITION \
-	MS_FILTER_METHOD(MSFilterPlayerInterface,8,int64_t)
+	MS_FILTER_METHOD(MSFilterPlayerInterface,8,int)
 
 #define MS_PLAYER_EOF \
 	MS_FILTER_EVENT_NO_ARG(MSFilterPlayerInterface,0)
